+++ src/main/java/com/example/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.example.controller;
 
-import com.example.core.user.service.UserService;
+import com.example.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
